# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Exemplo",
        "description": "Documentação da API de Exemplo",
        "contact": {
            "email": "contato@exemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/assunto": {
            "get": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Lista todos os assuntos",
                "description": "Retorna a lista de todos os assuntos",
                "operationId": "01f859a4a1dbbf1d412a123e8ab52782",
                "responses": {
                    "200": {
                        "description": "Lista de assuntos obtida com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Cria um novo assunto",
                "description": "Adiciona um novo assunto ao sistema",
                "operationId": "b8fa920ffd6d925308c1dae6ba968f38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string",
                                        "example": "Novo Assunto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assunto criado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/assunto/{id}/edit": {
            "get": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Exibe o formulário de edição de um assunto",
                "description": "Retorna o formulário para editar um assunto específico",
                "operationId": "d34018d49a64be5fe3b3df70b6a76c7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do assunto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulário de edição obtido com sucesso"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            }
        },
        "/assunto/{id}": {
            "put": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Atualiza um assunto",
                "description": "Atualiza os dados de um assunto existente",
                "operationId": "06070c494db2147f720c5ed371726292",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do assunto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição Atualizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assunto atualizado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assunto"
                ],
                "summary": "Exclui um assunto",
                "description": "Exclui logicamente um assunto (status = 0) se não estiver vinculado a um livro ativo",
                "operationId": "50c68d4b11cd2888e8093f42d565e190",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do assunto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assunto excluído com sucesso"
                    },
                    "400": {
                        "description": "Erro ao tentar excluir"
                    }
                }
            }
        },
        "/autor": {
            "get": {
                "tags": [
                    "Autor"
                ],
                "summary": "Lista todos os autores",
                "description": "Retorna a lista de todos os autores",
                "operationId": "293315e34d967ac4710c4d99ebaa0d7e",
                "responses": {
                    "200": {
                        "description": "Lista de autores obtida com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Autor"
                ],
                "summary": "Cria um novo autor",
                "description": "Adiciona um novo autor ao sistema",
                "operationId": "0986b88c2754eff0cd08a002d36ee5aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Novo Autor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor criado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/autor/{id}/edit": {
            "get": {
                "tags": [
                    "Autor"
                ],
                "summary": "Exibe o formulário de edição de um autor",
                "description": "Retorna o formulário para editar um autor específico",
                "operationId": "924dec31b54115c670c52db2f2c80e5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulário de edição obtido com sucesso"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            }
        },
        "/autor/{id}": {
            "put": {
                "tags": [
                    "Autor"
                ],
                "summary": "Atualiza um autor",
                "description": "Atualiza os dados de um autor existente",
                "operationId": "aacc0668c2b2bd3b0383e9d83cf95e5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Nome Atualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autor atualizado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Autor"
                ],
                "summary": "Exclui um autor",
                "description": "Exclui logicamente um autor (status = 0) se não estiver vinculado a um livro ativo",
                "operationId": "c7a7f0eea6605d70dd3e003039bd8c3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor excluído com sucesso"
                    },
                    "400": {
                        "description": "Erro ao tentar excluir"
                    }
                }
            }
        },
...
    },
...
}